+++ src/engine/scene.h
@@ -7,7 +7,7 @@ public:
     virtual ~i_scene() = default;
     virtual auto load()->void =0;
     virtual auto clean()->void =0;
-    virtual auto update(uint32_t delta_time)->void=0;
+    virtual auto update(float delta_time)->void=0;
     virtual auto render()->void = 0;
     virtual auto pause()->void = 0;
     virtual auto resume()->void = 0;
