+++ src/game/scene_main.h
@@ -13,7 +13,7 @@ public:
     auto clean() -> void override;
     auto pause() -> void override;
     auto resume() -> void override;
-    auto update(uint32_t delta_time) -> void override;
+    auto update(float delta_time) -> void override;
     auto render() -> void override;
     auto set_game(game* game) -> void override;
     auto handle_event(const input_state&) -> void override;
